Server: Make is_valid_http_request take a str instead of whole buffer
We only need to look at the http request header so we not worth passing
a reference to the buffer and then having to convert to a string in each
method. Better to convert use String::from_utf8_lossy once in at the top
level and pass that &str around.