Server: Switch server to use UTC time
This avoid displaying whatever time the docker container is set to.
Ideally, the time should be of the user but that will require some more
work. My first thought is that the request can include the users time
and then the time elapsed can be added but that doesn't seem very good.
The other option is to just remove timestamps in this primitive format
and see later on if TS or JS offers some nice features for displaying
local user time.